+++ app/Http/Controllers/API/UserAPIController.php
@@ -255,8 +255,15 @@ class UserAPIController extends Controller
         try {
             DB::beginTransaction();
                 $user = auth()->user();
+            if(auth()->getDefaultDriver() == "apiJwt")
+            {
                 $user->deleteDeviceToken($request->device_token);
                 auth()->logout();
+            }else{
+                if ($request->has('device_token')) {
+                    $user->setDeviceToken();
+                }
+            }
             DB::commit(); 
             return $this->sendResponse($user->name, 'User logout successfully');
